Refactor image model types and update imports; add ApiResponse type definition
