+++ app/models/ImageProps.ts
@@ -1,4 +1,4 @@
-import type { ImageDetails } from "../ImageDetails";
+import type { ImageDetails } from "./ImageDetails";
 
 type ImageCardProps = {
   image: ImageDetails;
