+++ app/ui/LoadingIndicator.tsx
@@ -6,12 +6,9 @@ import type { LoadingIndicatorProps } from "../models/UIComponentProps";
 /**
  * Renders a loading indicator using the BarLoader component from the react-spinners library.
  *
- * @component
- * @param {LoadingIndicatorProps} props - The component props.
- * @param {boolean} props.loading - Indicates whether the loading indicator should be displayed.
- * @param {string} [props.color] - The color of the loading indicator.
- * @param {number} [props.height] - The height of the loading indicator.
- * @returns {JSX.Element} The rendered LoadingIndicator component.
+ * @param props.loading - Indicates whether the loading indicator should be displayed.
+ * @param props.color - The color of the loading indicator.
+ * @param props.height - The height of the loading indicator.
  */
 const LoadingIndicator = ({
   color,
