+++ app/ui/image/ImageDetailsDisplay.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import type { ImageDetailsDisplayProps } from "@/app/models/image/ImageProps";
+import type { ImageDetailsDisplayProps } from "@/app/models/ImageProps";
 import Link from "next/link";
 
 /**
